REFACTOR: Consolidate all tensor network code in new submodule.
