+++ src/linear_algebra/LinearAlgebra.jl
@@ -2,9 +2,11 @@ module LinearAlgebra
 
 using Reexport
 
+include("TensorNetworks.jl")
 include("GLA.jl")
 include("StructuredGLA.jl")
 
+@reexport using .TensorNetworks
 @reexport using .GraphicalLinearAlgebra
 @reexport using .StructuredGraphicalLinearAlgebra
