+++ src/programs/RelationalPrograms.jl
@@ -2,8 +2,7 @@
 """
 module RelationalPrograms
 export RelationDiagram, UntypedRelationDiagram, TypedRelationDiagram,
-  @relation, @tensor_network, @eval_tensor_network,
-  parse_relation_diagram, parse_tensor_network, compile_tensor_expr
+  @relation, parse_relation_diagram
 
 using Compat
 using MLStyle: @match
@@ -66,9 +65,8 @@ function RelationDiagram{Name}(ports::AbstractVector{T};
   return d
 end
 
-
-# Relations
-###########
+# Relation macro
+################
 
 """ Construct an undirected wiring diagram using relation notation.
 
@@ -192,151 +190,6 @@ function parse_relation_call(call)
   else
     error("Relation call $call mixes named and unnamed arguments")
   end
-end
-
-# Tensor networks
-#################
-
-""" Construct an undirected wiring diagram using tensor notation.
-
-The tensor syntax is compatible with that used by packages like
-[TensorOperations.jl](https://github.com/Jutho/TensorOperations.jl) and
-[TensorCast.jl](https://github.com/mcabbott/TensorCast.jl). For example, the
-wiring diagram for composite of two matrices (or two binary relations) is
-constructed by
-
-```julia
-@tensor_network (i,j,k) C[i,k] := A[i,j] * B[j,k]
-```
-
-The leading context, or list of variables, may be omitted, in which case it is
-inferred from the variables used in the tensor expression. So, in this example,
-an equivalent macro call is
-
-```julia
-@tensor_network C[i,k] := A[i,j] * B[j,k]
-```
-
-See also: [`@eval_tensor_network`](@ref), the "inverse" to this macro.
-"""
-macro tensor_network(exprs...)
-  :(parse_tensor_network($((QuoteNode(expr) for expr in exprs)...)))
-end
-
-""" Parse an undirected wiring diagram from a tensor expression.
-
-For more information, see the corresponding macro [`@tensor_network`](@ref).
-"""
-function parse_tensor_network(context::Expr, expr::Expr)
-  all_vars = @match context begin
-    Expr(:tuple, args...) => collect(Symbol, args)
-    Expr(:vect, args...) => collect(Symbol, args)
-  end
-  parse_tensor_network(expr, all_vars=all_vars)
-end
-
-function parse_tensor_network(expr::Expr; all_vars=nothing)
-  # Parse tensor expression.
-  (outer_name, outer_vars), body = @match expr begin
-    Expr(:(=), outer, body) => (parse_tensor_term(outer), body)
-    Expr(:(:=), outer, body) => (parse_tensor_term(outer), body)
-    _ => error("Tensor expression $expr must be an assignment, either = or :=")
-  end
-  names_and_vars = map(parse_tensor_term, @match body begin
-    Expr(:call, :(*), args...) => args
-    1 => [] # No terms
-    arg => [arg] # One term
-  end)
-
-  # Check compatibility of used variables with declared variables.
-  used_vars = unique!(reduce(vcat, ([[outer_vars]; last.(names_and_vars)])))
-  if isnothing(all_vars)
-    all_vars = sort!(used_vars)
-  else
-    used_vars ⊆ all_vars || error("One of variables $used_vars is not declared")
-  end
-
-  # Construct the undirected wiring diagram.
-  d = RelationDiagram{Symbol}(length(outer_vars))
-  add_junctions!(d, length(all_vars), variable=all_vars)
-  set_junction!(d, ports(d, outer=true),
-                incident(d, outer_vars, :variable), outer=true)
-  for (name, vars) in names_and_vars
-    box = add_box!(d, length(vars), name=name)
-    set_junction!(d, ports(d, box), incident(d, vars, :variable))
-  end
-  return d
-end
-
-function parse_tensor_term(expr)
-  @match expr begin
-    Expr(:ref, name::Symbol, args...) => (name, collect(Symbol, args))
-    name::Symbol => (name, Symbol[]) # Scalar
-    _ => error("Invalid syntax in term $expr in tensor expression")
-  end
-end
-
-""" Evaluate a tensor network using another macro.
-
-This macro takes two arguments: an undirected wiring diagram and a macro call
-supporting the tensor contraction notation that is de facto standard among Julia
-tensor packages. For example, to evaluate a tensor network using the `@tullio`
-macro, use:
-
-```julia
-A = @eval_tensor_network diagram @tullio
-```
-
-The following macros should work:
-
-- `@tensor` from
-  [TensorOperations.jl](https://github.com/Jutho/TensorOperations.jl).
-- `@tullio` from [Tullio.jl](https://github.com/mcabbott/Tullio.jl)
-- `@einsum` from [Einsum.jl](https://github.com/ahwillia/Einsum.jl)
-- `@ein` from [OMEinsum.jl](https://github.com/under-Peter/OMEinsum.jl)
-
-However, the macros `@cast` and `@reduce` from
-[TensorCast.jl](https://github.com/mcabbott/TensorCast.jl) will *not* work
-because they do not support implicit summation.
-
-See also: [`@tensor_network`](@ref), the "inverse" to this macro.
-"""
-macro eval_tensor_network(diagram, tensor_macro)
-  # XXX: We cannot use `GeneralizedGenerated.mk_function` here because packages
-  # like Tullio generate code with type parameters, which GG does not allow.
-  compile_expr = :(compile_tensor_expr($(esc(diagram)),
-    assign_op=:(:=), assign_name=gensym("out")))
-  Expr(:call, esc(:eval),
-       :(_eval_tensor_network($compile_expr, $(QuoteNode(tensor_macro)))))
-end
-function _eval_tensor_network(tensor_expr, macro_expr)
-  @match macro_expr begin
-    Expr(:macrocall, args...) => Expr(:macrocall, args..., tensor_expr)
-    _ => error("Expression $macro_expr is not a macro call")
-  end
-end
-
-""" Generate tensor expression from undirected wiring diagram.
-
-This function is used to implement [`@eval_tensor_network`](@ref) but may be
-useful in its own right.
-"""
-function compile_tensor_expr(d::UndirectedWiringDiagram;
-    assign_op::Symbol=:(=), assign_name::Symbol=:out)
-  vars = j -> subpart(d, j, :variable)
-  outer_vars = vars(junction(d, ports(d, outer=true), outer=true))
-  terms = map(boxes(d)) do box
-    ref_expr(subpart(d, box, :name), vars(junction(d, ports(d, box))))
-  end
-  lhs = ref_expr(assign_name, outer_vars)
-  rhs = if isempty(terms); 1
-    elseif length(terms) == 1; first(terms)
-    else Expr(:call, :(*), terms...) end
-  Expr(assign_op, lhs, rhs)
-end
-
-function ref_expr(name::Symbol, vars)
-  isempty(vars) ? name : Expr(:ref, name, vars...)
 end
 
 end
