+++ test/linear_algebra/LinearAlgebra.jl
@@ -1,5 +1,9 @@
 using Test
 
+@testset "TensorNetworks" begin
+  include("TensorNetworks.jl")
+end
+
 @testset "GraphicalLinearAlgebra" begin
   include("GLA.jl")
   include("StructuredGLA.jl")
